requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt ./config/requirements.in
#
click==6.7                # via pip-tools
dj-database-url==0.4.2
django-anymail==0.8
django-crispy-forms==1.6.1
django-extensions==1.7.5
django-filter==1.0.1
django-html5==1.0.0
django-nose==1.4.4
django-settings-export==1.2.1
django==1.10.5
djangorestframework==3.6.2
factory-boy==2.8.1
faker==0.7.7              # via factory-boy
first==2.0.1              # via pip-tools
gunicorn==19.6.0
inflection==0.3.1
nose==1.3.7               # via django-nose
numpy==1.12.0             # via pandas
pandas==0.19.2
pip-tools==1.8.0
psycopg2==2.6.2
python-dateutil==2.6.0    # via faker, pandas
pytz==2016.10             # via pandas
requests==2.13.0          # via django-anymail
six==1.10.0               # via django-anymail, django-extensions, faker, pip-tools, python-dateutil
werkzeug==0.11.15
whitenoise==3.3.0
